requirements.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==2.6.11
Pillow==5.2.0
psycopg2==2.7.7
PyJWT==1.6.4
pytz==2018.5
